Moved function to utils
